// src/models/anime.ts
export interface Anime {
  japanese_name: string;
  name: string;
  rank: number;
  rating: number;
  related_anime: string[];
  release_season: string;
  release_year: number;
  studio: string;
  tags: string[];
  type: string;
}

export interface AnimeJson {
  Japanese_name: string;
  Name: string;
  Rank: number;
  Rating: number;
  Related_anime: string[];
  Release_season: string;
  Release_year: number;
  Studio: string;
  Tags: string;
  Type: string;
}